nyc-bikes-pipeline/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
boto3==1.34.0
psycopg2-binary==2.9.9
pandas==2.1.3
mage-ai==0.9.70
streamlit==1.28.0
plotly==5.17.0
requests==2.31.0